requirements.txt
matplotlib
opencv-python
Pillow
ruamel.yaml